requirements.txt
annotated-types==0.7.0
anyio==3.7.1
blinker==1.8.2
certifi==2025.1.31
click==8.1.8
distro==1.9.0
elastic-transport==8.17.0
elasticsearch==8.17.1
Flask==2.2.5
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
MarkupSafe==2.1.5
openai==1.62.0
packaging==24.2
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3<2.0.0
Werkzeug==2.2.3
Flask-Cors==3.0.10